Refactor S3 store & support upload retry
* Upload failures for S3 will now retry
* S3 tcp connections now counted as with file descriptor limits
* Significantly refactored how uploads happen in S3 store

+++ nativelink-util/src/buf_channel.rs
@@ -243,8 +243,8 @@ impl DropCloserReadHalf {
     /// Sets the maximum size of the recent_data buffer. If the number of bytes
     /// received exceeds this size, the recent_data buffer will be cleared and
     /// no longer populated.
-    pub fn set_max_recent_data_size(&mut self, size: u64) {
-        self.max_recent_data_size = size;
+    pub fn set_max_recent_data_size(&mut self, size: usize) {
+        self.max_recent_data_size = size as u64;
     }
 
     /// Attempts to reset the stream to before any data was received. This will
